+++ ent/heuristic/heuristic.go
@@ -29,28 +29,28 @@ func exploreNameCandidate(ts []token.TokenSN, d *dict.Dictionary) bool {
 
 	u := ts[0]
 
-	if u.Features().UninomialDict == dict.WhiteUninomial {
+	if u.Features().UninomialDict == dict.InUninomial {
 		u.SetDecision(token.Uninomial)
 		return true
 	}
 
-	if u.Features().UninomialDict == dict.GreyUninomial {
+	if u.Features().UninomialDict == dict.InAmbigUninomial {
 		u.SetDecision(token.PossibleUninomial)
 		return true
 	}
 
 	if u.Indices().Species == 0 {
-		if u.Features().UninomialDict == dict.WhiteGenus {
+		if u.Features().UninomialDict == dict.InGenus {
 			u.SetDecision(token.Uninomial)
 			return true
 		}
-		if u.Features().UninomialDict == dict.GreyGenus {
+		if u.Features().UninomialDict == dict.InAmbigGenus {
 			u.SetDecision(token.PossibleUninomial)
 			return true
 		}
 	}
 
-	if u.Features().UninomialDict == dict.BlackUninomial {
+	if u.Features().UninomialDict == dict.NotInUninomial {
 		return false
 	}
 
@@ -68,8 +68,8 @@ func exploreNameCandidate(ts []token.TokenSN, d *dict.Dictionary) bool {
 
 func checkAsSpecies(t token.TokenSN) bool {
 	if !t.Features().IsCapitalized &&
-		(t.Features().SpeciesDict == dict.WhiteSpecies ||
-			t.Features().SpeciesDict == dict.GreySpecies) {
+		(t.Features().SpeciesDict == dict.InSpecies ||
+			t.Features().SpeciesDict == dict.InAmbigSpecies) {
 		return true
 	}
 	return false
@@ -79,28 +79,28 @@ func checkAsGenusSpecies(ts []token.TokenSN, d *dict.Dictionary) bool {
 	g := ts[0]
 	s := ts[g.Indices().Species]
 	if !checkAsSpecies(s) {
-		if g.Features().UninomialDict == dict.WhiteGenus {
+		if g.Features().UninomialDict == dict.InGenus {
 			g.SetDecision(token.Uninomial)
 			return true
 		}
-		if g.Features().UninomialDict == dict.GreyGenus {
+		if g.Features().UninomialDict == dict.InAmbigGenus {
 			g.SetDecision(token.PossibleUninomial)
 			return true
 		}
 		return false
 	}
 
-	if g.Features().UninomialDict == dict.WhiteGenus {
+	if g.Features().UninomialDict == dict.InGenus {
 		g.SetDecision(token.Binomial)
 		return true
 	}
 
-	if checkGreyGeneraSp(g, s, d) {
+	if checkInAmbigGeneraSp(g, s, d) {
 		g.SetDecision(token.Binomial)
 		return true
 	}
 
-	if s.Features().SpeciesDict == dict.WhiteSpecies &&
+	if s.Features().SpeciesDict == dict.InSpecies &&
 		!s.Features().IsCapitalized {
 		g.SetDecision(token.PossibleBinomial)
 		return true
@@ -108,26 +108,26 @@ func checkAsGenusSpecies(ts []token.TokenSN, d *dict.Dictionary) bool {
 	return false
 }
 
-func checkGreyGeneraSp(
+func checkInAmbigGeneraSp(
 	g token.TokenSN,
 	s token.TokenSN,
 	d *dict.Dictionary,
 ) bool {
 	sp := fmt.Sprintf("%s %s", g.Cleaned(), s.Cleaned())
-	if _, ok := d.GreyGeneraSp[sp]; ok {
-		g.Features().GenSpGreyDict += 1
+	if _, ok := d.InAmbigGeneraSp[sp]; ok {
+		g.Features().GenSpInAmbigDict += 1
 		return true
 	}
 	return false
 }
 
-func checkGreyGeneraIsp(
+func checkInAmbigGeneraIsp(
 	g, s, isp token.TokenSN,
 	d *dict.Dictionary,
 ) bool {
 	name := fmt.Sprintf("%s %s %s", g.Cleaned(), s.Cleaned(), isp.Cleaned())
-	if _, ok := d.GreyGeneraSp[name]; ok {
-		g.Features().GenSpGreyDict += 1
+	if _, ok := d.InAmbigGeneraSp[name]; ok {
+		g.Features().GenSpInAmbigDict += 1
 		return true
 	}
 	return false
@@ -142,7 +142,7 @@ func checkInfraspecies(ts []token.TokenSN, d *dict.Dictionary) {
 	s := ts[ts[0].Indices().Species]
 	isp := ts[i]
 
-	if checkGreyGeneraIsp(g, s, isp, d) || checkAsSpecies(ts[i]) {
+	if checkInAmbigGeneraIsp(g, s, isp, d) || checkAsSpecies(ts[i]) {
 		ts[0].SetDecision(token.Trinomial)
 	}
 }